two executables working now
