+++ src/mbot_omni.c
@@ -1,1 +1,57 @@
+/**
+ * This file is the main executable for the MBot OMNI firmware.
+ */
 #include "mbot_omni.h"
+
+
+static bool running = false;
+mbot_params_t params;
+
+int main()
+{   
+    running = false;
+    mbot_init_pico();
+    mbot_init_hardware();
+    mbot_init_comms();
+    mbot_read_fram(0, sizeof(params), &params);
+    
+    //Check also that define drive type is same as FRAM drive type
+    int validate_status = validate_FRAM_data(&params);
+    if (validate_status < 0)
+    {
+        printf("Failed to validate FRAM Data! Error code: %d\n", validate_status);
+        return -1;
+    }
+
+    if(params.robot_type != MBOT_DRIVE_TYPE){
+        printf("#define type is not equal to calibration type!\n");
+        return -1;
+    }
+
+    sleep_ms(3000);
+    print_mbot_params(&params);
+    printf("Starting MBot Loop...\n");
+    repeating_timer_t loop_timer;
+    add_repeating_timer_ms(MAIN_LOOP_PERIOD * 1000, mbot_loop, NULL, &loop_timer); // 1000x to convert to ms
+    printf("Done Booting Up!\n");
+    running = true;
+    uint16_t counter = 0;
+    
+    while(running){
+        // Heartbeat
+        if(!(counter % 5)){
+            gpio_put(LED_PIN, 1);
+        }
+        else if(!(counter % 7)){
+            gpio_put(LED_PIN, 1);
+            counter = 0;
+        }
+        else{
+            gpio_put(LED_PIN, 0);
+        }
+        // Print State
+        mbot_print_state(mbot_imu, mbot_encoders, mbot_odometry, mbot_motor_vel);
+        sleep_ms(200);  
+        counter++;
+    }
+}
